IAT-216: - Updated the README to be accurate - Added a new script in pack.json, it allows running the frontend with the prod profile active - Renamed component navbar to header
