+++ package.json
@@ -8,6 +8,7 @@
     "ng-build": "ng build",
     "ng-build-prod": "ng build -prod",
     "start": "ng serve  --proxy-config proxy.conf.json",
+    "start-prod": "ng serve  --proxy-config proxy.conf.json --prod",
     "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check",
     "lint-on-ci": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check --formatters-dir node_modules/tslint-teamcity-reporter/ --format TSHintTeamcity file.ts && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check --formatters-dir node_modules/tslint-teamcity-reporter/ --format TSHintTeamcity file.ts",
     "test": "ng test",
