+++ src/app/app.component.html
@@ -1,2 +1,2 @@
-<app-navbar></app-navbar>
+<app-header></app-header>
 <router-outlet></router-outlet>
