+++ src/app/header/header-routes.config.ts
@@ -1,4 +1,4 @@
-import { MenuType, RouteInfo } from './navbar.metadata';
+import { MenuType, RouteInfo } from './header.metadata';
 
 export const ROUTES: RouteInfo[] = [
   { path: '',         title: 'Angular2 Bootstrap4 Navbar',  menuType: MenuType.BRAND },
