+++ src/app/header/header.component.html
@@ -1,0 +1,19 @@
+
+<nav class="navbar navbar-toggleable-md navbar-light" >
+
+  <a class="navbar-brand" href="#">
+    <img src="/assets/images/brand_sbac.png" width="50px" height="50px"  class="d-inline-block align-top" alt=""/>
+  </a>
+
+  <div class="collapse navbar-collapse" id="navbarText">
+    <ul class="navbar-nav mr-auto">
+      <li class="nav-item">Smarter App - Item Authoring Tool</li>
+    </ul>
+    <ul class="navbar-nav">
+      <li class="nav-item align-content-end">
+        <a class="btn btn-outline-secondary btn-sm" href="/saml/logout/">Log out</a>
+      </li>
+    </ul>
+  </div>
+
+</nav>
