+++ src/app/header/header.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { ROUTES } from './navbar-routes.config';
-import { MenuType } from './navbar.metadata';
+import { ROUTES } from './header-routes.config';
+import { MenuType } from './header.metadata';
 
 @Component({
-  selector: 'app-navbar',
-  templateUrl: 'navbar.component.html',
-  styleUrls: [ 'navbar.component.scss' ]
+  selector: 'app-header',
+  templateUrl: 'header.component.html',
+  styleUrls: [ 'header.component.scss' ]
 })
-export class NavbarComponent implements OnInit {
+export class HeaderComponent implements OnInit {
   public menuItems: any[];
   public brandMenu: any;
   isCollapsed = true;
