+++ src/app/home/home.component.html
@@ -1,6 +1,6 @@
 <header>
   <div class="jumbotron jumbotron-fluid" id="banner">
-    <div class="parallax text-center" style="background-image: url(/assets/images/shield-large.png);">
+    <div class="parallax text-center">
       <div class="parallax-pattern-overlay">
         <div class="container text-center" id="size-banner">
           <h2 class="display-2">Smarter Balanced</h2>
