--- src/app/navbar/index.ts
@@ -1,1 +1,0 @@
-export * from './navbar.module';
