--- src/app/navbar/navbar.component.html
@@ -1,32 +1,0 @@
-<nav class="navbar navbar-toggleable-md navbar-light bg-faded">
-
-  <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-
-  <a class="navbar-brand" href="#"><img src="/assets/images/logo_sbac.jpg"/></a>
-
-  <div class="collapse navbar-collapse" id="navbarSupportedContent">
-    <ul class="navbar-nav mr-auto">
-      <li class="nav-item active">
-        <a class="nav-link" href="#">Link 1 <span class="sr-only">(current)</span></a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link" href="#">Link 2</a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link disabled" href="#">Disabled</a>
-      </li>
-    </ul>
-    <ul class="navbar-nav">
-      <li class="nav-item align-content-end">
-        <a class="nav-link" href="/saml/logout/">Logout</a>
-      </li>
-    </ul>
-
-    <!--<form class="form-inline">-->
-      <!--<input class="form-control mr-sm-2" type="text" placeholder="Search">-->
-      <!--<button class="btn btn-outline-success " type="submit">Search</button>-->
-    <!--</form>-->
-  </div>
-</nav>
